+++ minio_storage/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class DjangoStoragesConfig(AppConfig):
+    name = "minio_storage"