+++ minio_storage/policy.py
@@ -4,6 +4,7 @@ import typing as T
 
 
 class Policy(enum.Enum):
+    none = "NONE"
     get = "GET_ONLY"
     read = "READ_ONLY"
     write = "WRITE_ONLY"
@@ -17,11 +18,16 @@ class Policy(enum.Enum):
             Policy.read: _read,
             Policy.write: _write,
             Policy.read_write: _read_write,
+            Policy.none: _none,
         }
         pol = policies[self](bucket_name)
         if json_encode:
             return json.dumps(pol)
         return pol
+
+
+def _none(bucket_name: str) -> T.Dict:
+    return {"Version": "2012-10-17", "Statement": []}
 
 
 def _get(bucket_name: str) -> T.Dict:
@@ -31,15 +37,9 @@ def _get(bucket_name: str) -> T.Dict:
             {
                 "Effect": "Allow",
                 "Principal": {"AWS": ["*"]},
-                "Action": ["s3:GetBucketLocation"],
-                "Resource": [f"arn:aws:s3:::{bucket_name}"],
-            },
-            {
-                "Effect": "Allow",
-                "Principal": {"AWS": ["*"]},
                 "Action": ["s3:GetObject"],
                 "Resource": [f"arn:aws:s3:::{bucket_name}/*"],
-            },
+            }
         ],
     }